move demos to ui folder
